+++ @xen-orchestra/backups/_backupJob/RemoteTimeoutError.js
@@ -1,0 +1,7 @@
+'use strict'
+exports.RemoteTimeoutError = class RemoteTimeoutError extends Error {
+  constructor(remoteId) {
+    super('timeout while getting the remote ' + remoteId)
+    this.remoteId = remoteId
+  }
+}
