+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_MixinXapiWriter.js
@@ -1,8 +1,8 @@
 'use strict'
 
-const { Task } = require('../Task')
+const { Task } = require('../../../Task')
 const assert = require('node:assert/strict')
-const { HealthCheckVmBackup } = require('../HealthCheckVmBackup')
+const { HealthCheckVmBackup } = require('../../../HealthCheckVmBackup')
 
 function extractOpaqueRef(str) {
   const OPAQUE_REF_RE = /OpaqueRef:[0-9a-z-]+/
@@ -12,8 +12,8 @@ function extractOpaqueRef(str) {
   }
   return matches[0]
 }
-exports.MixinReplicationWriter = (BaseClass = Object) =>
-  class MixinReplicationWriter extends BaseClass {
+exports.MixinXapiWriter = (BaseClass = Object) =>
+  class MixinXapiWriter extends BaseClass {
     constructor({ sr, ...rest }) {
       super(rest)
