+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_checkVhd.js
