+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_getOldEntries.js
