+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_listReplicatedVms.js
