+++ @xen-orchestra/backups/_backupJob/VmBackup/writers/_packUuid.js
