+++ @xen-orchestra/backups/_backupJob/createStreamThrottle.js
