+++ @xen-orchestra/backups/_backupJob/getAdapterByRemote.js
@@ -1,0 +1,9 @@
+'use strict'
+
+exports.getAdaptersByRemote = adapters => {
+  const adaptersByRemote = {}
+  adapters.forEach(({ adapter, remoteId }) => {
+    adaptersByRemote[remoteId] = adapter
+  })
+  return adaptersByRemote
+}
