+++ packages/xo-web/src/xo-app/backup/new/new-schedule.js
@@ -177,7 +177,7 @@ const New = decorate([
             </div>
           )}
         </FormGroup>
-        {modes.isDelta && (
+        {modes.isIncremental && (
           <FormGroup>
             <label>
               <strong>{_('forceFullBackup')}</strong>{' '}
